Changed how components are used and added some configurable options
Changes were made to allow for onClick handlers to be passed to the action buttons.
Changes were also made to allow for smoother transitions.
Props were added to allow buttons to be revealed on hover (optional).
A style tag can be used to allow for easy positioning within the element that contains
the component.